Exclude shaded variant of Kotlin's ModuleDescriptorImpl
Issue #146 identified a quasar instrumentation problem that causes a
`VerifyError` in Kotlin's `ModuleDescriptorImpl`. The fix for #146
explicitly excludes this class by its fully qualified name
`kotlin.reflect.jvm.internal.impl.descriptors.impl.ModuleDescriptorImpl`,
and this works, but this same class is shaded/repackaged as
`org.jetbrains.kotlin.descriptors.impl.ModuleDescriptorImpl` within the
kotlin-compiler[-embeddable] jars.

This fix adds the latter to the set of exclusions so the workaround
applies to both cases.