Bumps versions for Java 9 compatibility.
We bump the shadow plugin to 2.0.2 in order to get version 6 of ASM,
which is now compatible with Java 9.  The reason we don't bump it to
2.0.3 is because 2.0.3 uses ASM version 6.1, which has a bug (fixed in
6.1.1) in `ClassReader.readStream`.  It allocates a data buffer with
size equal to the available bytes in the input stream.  However, in
practice, the input stream is a `ZipFile` `BoundedInputStream`, which
doesn't implement the `available` method, so the size of the data buffer
is 0, causing an infinite loop.  There isn't presently a later version
of the shadow plugin that uses ASM 6.1.1.